+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [UNRELEASED]
 ### Added
 - [LIBRARY] Add rendering hooks.
+- [INTERNAL] Add tests for hooks and renderers
 
 
 ## [2.0.106] - 20200329